\name{Printing}
\alias{Printing}
\title{Printing functions for objects of class latex_symb}
\usage{
    a <- latex_symb("\\alpha")
    print(a)
    as.character(a)
}
\arguments{
    \item{x}{An object of class latex_symb.}
}
\description{
    Print the object in a format suitable for inclusion in a LaTeX document.
}
