% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{braces}
\alias{braces}
\title{Function that creates a latex_symb object consisting of its argument enclosed in braces}
\usage{
braces(expr)
}
\arguments{
\item{expr}{An object}
}
\value{
A latex_symb object. Its representation is the argument enclosed in braces.
}
\description{
Function that creates a latex_symb object consisting of its argument enclosed in braces
}
