% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{il}
\alias{il}
\title{Function that wraps a `latex_symb` object in dollar signs to be interpreted as inline math by `LaTeX`.}
\usage{
il(expr)
}
\arguments{
\item{expr}{An object}
}
\value{
A string. It is the representation of `expr`, wrapped by dollar signs.
}
\description{
Function that wraps a `latex_symb` object in dollar signs to be interpreted as inline math by `LaTeX`.
}
\examples{
e1 <- lsymb("\\\\alpha")
il(e1)
}
