% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lsymb}
\alias{lsymb}
\title{Function that creates a `latex_symb` object}
\usage{
lsymb(...)
}
\arguments{
\item{...}{One or more R objects, which get transformed into characters and then pasted to create the `LaTeX` code for the symbol.}
}
\value{
An object of the S3 class `latex_symb`.
}
\description{
Function that creates a `latex_symb` object
}
\examples{
al <- lsymb("\\\\alpha")
al$repr
}
