% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.latex_symb}
\alias{print.latex_symb}
\title{Prints the `LaTeX` representation of a `latex_symb` object}
\usage{
\method{print}{latex_symb}(x, ...)
}
\arguments{
\item{x}{A `latex_symb` object}

\item{...}{Other arguments}
}
\value{
The object's representation
}
\description{
Prints the `LaTeX` representation of a `latex_symb` object
}
