% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pths}
\alias{pths}
\title{Function that creates a `latex_symb` object consisting of its argument enclosed in parentheses}
\usage{
pths(expr)
}
\arguments{
\item{expr}{An object}
}
\value{
A `latex_symb` object. Its representation is the argument enclosed in parentheses.
}
\description{
Function that creates a `latex_symb` object consisting of its argument enclosed in parentheses
}
\examples{
e1 <- lsymb("\\\\alpha")
pths(e1)
}
