% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{/}
\alias{/}
\title{Masking the division operator to work with latex_symb objects}
\usage{
e1 / e2
}
\arguments{
\item{e1}{An object}

\item{e2}{An object}
}
\value{
A latex_symb object if one of the arguments is a latex_symb object. Its representation is the latex operator frac, with the representation of the first argument as the numerator and the representation of the second argument as the denominator. Otherwise, the base division operator is called.
}
\description{
Masking the division operator to work with latex_symb objects
}
