\name{common}
\docType{data}
\alias{common}
\title{Common latex_symb objects}
\description{
    A collection of common latex_symb objects.
}
\format{
  An RData file containing:
  \describe{
    \item{i}{latex_symb object whose repr is "i"}
    \item{j}{latex_symb object whose repr is "j"}
    \item{k}{latex_symb object whose repr is "k"}
    \item{l}{latex_symb object whose repr is "l"}
    \item{m}{latex_symb object whose repr is "m"}
    \item{n}{latex_symb object whose repr is "n"}
    \item{x}{latex_symb object whose repr is "x"}
    \item{y}{latex_symb object whose repr is "y"}
    \item{z}{latex_symb object whose repr is "z"}
    \item{f}{latex_symb object whose repr is "f"}
    \item{g}{latex_symb object whose repr is "g"}
    \item{h}{latex_symb object whose repr is "h"}
    \item{al}{latex_symb object whose repr is "\\\\alpha"}
    \item{be}{latex_symb object whose repr is "\\\\beta"}
    \item{ga}{latex_symb object whose repr is "\\\\gamma"}
    \item{de}{latex_symb object whose repr is "\\\\delta"}
    \item{ep}{latex_symb object whose repr is "\\\\epsilon"}
    \item{ze}{latex_symb object whose repr is "\\\\zeta"}
    \item{et}{latex_symb object whose repr is "\\\\eta"}
    \item{th}{latex_symb object whose repr is "\\\\theta"}
    \item{io}{latex_symb object whose repr is "\\\\iota"}
    \item{ka}{latex_symb object whose repr is "\\\\kappa"}
    \item{la}{latex_symb object whose repr is "\\\\lambda"}
    \item{mu}{latex_symb object whose repr is "\\\\mu"}
    \item{nu}{latex_symb object whose repr is "\\\\nu"}
    \item{xi}{latex_symb object whose repr is "\\\\xi"}
    \item{om}{latex_symb object whose repr is "\\\\omicron"}
    \item{pi.l}{latex_symb object whose repr is "\\\\pi"}
    \item{rh}{latex_symb object whose repr is "\\\\rho"}
    \item{si}{latex_symb object whose repr is "\\\\sigma"}
    \item{ta}{latex_symb object whose repr is "\\\\tau"}
    \item{up}{latex_symb object whose repr is "\\\\upsilon"}
    \item{ph}{latex_symb object whose repr is "\\\\phi"}
    \item{ch}{latex_symb object whose repr is "\\\\chi"}
    \item{ps}{latex_symb object whose repr is "\\\\psi"}
    \item{om}{latex_symb object whose repr is "\\\\omega"}
    \item{des}{latex_symb object whose repr is ":"}
    \item{eq}{latex_symb object whose repr is "="}
    \item{neq}{latex_symb object whose repr is "\\\\neq"}
    \item{lt}{latex_symb object whose repr is "<"}
    \item{gt}{latex_symb object whose repr is ">"}
    \item{leq}{latex_symb object whose repr is "\\\\leq"}
    \item{geq}{latex_symb object whose repr is "\\\\geq"}
    \item{bgs}{latex_symb object whose repr is "\\\\in"}
    \item{mapsto}{latex_symb object whose repr is "\\\\mapsto"}
    \item{to}{latex_symb object whose repr is "\\\\rightarrow"}
    \item{ldots}{latex_symb object whose repr is "\\\\ldots"}
    \item{Reals}{latex_symb object whose repr is "\\\\mathbb\{R\}"}
    \item{Nats}{latex_symb object whose repr is "\\\\mathbb\{N\}"}
    \item{Ints}{latex_symb object whose repr is "\\\\mathbb\{Z\}"}
    \item{Rats}{latex_symb object whose repr is "\\\\mathbb\{Q\}"}
    \item{Comps}{latex_symb object whose repr is "\\\\mathbb\{C\}"}
    \item{indic}{latex_symb object whose repr is "\\\\mathbb\{1\}"}
    \item{infty}{latex_symb object whose repr is "\\\\infty"}
  }
}