% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{latex_symb}
\alias{latex_symb}
\title{Function that creates a latex_symb object}
\usage{
latex_symb(repr)
}
\arguments{
\item{repr}{A character string representing the latex expression}
}
\value{
A latex_symb object
}
\description{
Function that creates a latex_symb object
}
\examples{
al <- latex_symb("\\\\alpha")
}
