% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.latex_symb}
\alias{print.latex_symb}
\title{Adaptation of the generic print function for latex_symb objects}
\usage{
\method{print}{latex_symb}(x, ...)
}
\arguments{
\item{x}{A latex_symb object}

\item{...}{Additional arguments}
}
\description{
Adaptation of the generic print function for latex_symb objects
}
